-nohull2 also disables hull 3, for even less clipnodes. (WIP)
The BSP model needs to be manually updated to redirect hull 3 references to hull 1, so that you don't fall the level when you crouch. The edit_hulls.py will show which byte needs to be edited or something, but not do it for you.

TODO: add a separate -nohull3 option for this, and automatically redirect hull 3 to hull 1 in the bsp.